platform/pyproject.toml
[project]
name = "personal-platform"
requires-python = "~=3.13.9"
version = "0.1.0"
dependencies = [
  "anthropic>=0.64.0",
  "anyio>=4.10.0",
  "beautifulsoup4>=4.13.5",
  "camelot-py>=1.0.9",
  "claude-code-sdk>=0.0.20",
  "duckdb>=1.3.2",
  "elasticsearch>=9.1.0",
  "firecrawl>=3.4.0",
  "genson>=1.3.0",
  "httpx>=0.28.1",
  "humanize>=4.13.0",
  "jsonschema>=4.25.1",
  "lxml>=6.0.1",
  "markdownify>=1.2.0",
  "matplotlib>=3.10.5",
  "nest-asyncio>=1.6.0",
  "openai>=2.7.0",
  "opencv-python>=4.12.0.88",
  "pandas>=2.3.2",
  "pdfplumber>=0.11.7",
  "pillow>=11.3.0",
  "playwright>=1.54.0",
  "polars>=1.32.3",
  "prance[osv]>=25.4.8.0",
  "pyjwt[crypto]>=2.10.1",
  "prompt-toolkit>=3.0.52",
  "pydantic>=2.11.7",
  "pymupdf>=1.26.4",
  "pymysql>=1.1.2",
  "pyyaml>=6.0.2",
  "requests>=2.32.5",
  "urllib3>=2.5.0",
  "rich>=14.1.0",
  "sqlfluff>=3.4.2",
  "typer>=0.19.2",
  "logfire[requests,httpx]>=4.14.2",
  "graphql-core>=3.2.6",
  "pypandoc>=1.15"
]

[dependency-groups]
dev = [
  "ruff>=0.12.10",
  "pytest>=8.4.1",
  "marimo[recommended]>=0.15.0",
  "watchdog>=6.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["./src"]
testpaths = ["src", "tests"]
addopts = "--capture=no"

[tool.ruff]
line-length = 160
target-version = "py313"

[tool.ruff.lint]
select = [
  "E", # pycodestyle
  "F", # pyflakes
  "I", # isort
  "B", # flake8-bugbear
]
extend-ignore = [
  # Skip unused-variable warning, Marimo will generate "unused" variables that are actually used
  "F841",
]

[tool.marimo.ai]
rules = "- prefer polars over pandas\n- make charts using altair"

[tool.marimo.save]
autosave = "after_delay"
autosave_delay = 1000

[tool.marimo.package_management]
manager = "uv"

[tool.marimo.runtime]
pythonpath = ["./src"]
auto_instantiate = false
